client: Move types to version specific module
Currently we have a bit of confusion around types that are defined and
used by the various client version modules.

- Some currently don't change e.g., `Input`
- Some are due to change in v29 e.g., `TemplateRequest`
- Some change already e.g. `AddressVersion`

And to make matters worse, I did 'custom' re-exports in `node` which is
very confusing when one is reading code in `client`.

Put all `client` defined types in the version module they first appear
in and re-export them as we do for types from `types`. Then use the type
naked (without qualifying the path).
