+++ client/src/client_sync/v17/wallet.rs
@@ -446,7 +446,7 @@ macro_rules! impl_client_v17__walletcreatefundedpsbt {
         impl Client {
             pub fn wallet_create_funded_psbt(
                 &self,
-                inputs: Vec<$crate::client_sync::WalletCreateFundedPsbtInput>,
+                inputs: Vec<WalletCreateFundedPsbtInput>,
                 outputs: Vec<BTreeMap<Address, Amount>>,
             ) -> Result<WalletCreateFundedPsbt> {
                 self.call("walletcreatefundedpsbt", &[into_json(inputs)?, into_json(outputs)?])
