+++ client/src/client_sync/v18/mod.rs
@@ -18,10 +18,10 @@ use crate::types::v18::*;
 
 #[rustfmt::skip]                // Keep public re-exports separate.
 pub use crate::client_sync::{
-    v17::{AddressType, TemplateRequest, TemplateRules},
-    WalletCreateFundedPsbtInput
+    v17::{AddressType, Input, Output, TemplateRequest, TemplateRules, WalletCreateFundedPsbtInput},
 };
 
+// This publicly re-exports `Client`.
 crate::define_jsonrpc_minreq_client!("v18");
 crate::impl_client_check_expected_server_version!({ [180100] });
