+++ client/src/client_sync/v23/mod.rs
@@ -19,8 +19,7 @@ use crate::types::v23::*;
 
 #[rustfmt::skip]                // Keep public re-exports separate.
 pub use crate::client_sync::{
-    v17::{TemplateRequest, TemplateRules},
-    WalletCreateFundedPsbtInput
+    v17::{Input, Output, TemplateRequest, TemplateRules, WalletCreateFundedPsbtInput},
 };
 
 crate::define_jsonrpc_minreq_client!("v23");
